Create postnl uploads subdirectory

+++ src/Order/Base.php
@@ -459,7 +459,9 @@ abstract class Base {
 					$filepath   = trailingslashit( POSTNL_UPLOADS_DIR ) . $filename;
 
 					$test     = base64_decode( $label_contents['Content'] );
-					$file_ret = file_put_contents( $filepath, $test );
+					if ( wp_mkdir_p( POSTNL_UPLOADS_DIR ) && ! file_exists( $filepath ) ) {
+						$file_ret = file_put_contents( $filepath, $test );
+					}
 
 					$labels[] = array(
 						'type'     => $label_type,
